Extract createSecureDocumentBuilder into build logic basics
Build logic parses XML in several places and cannot use
org.gradle.internal.xml.XmlFactories for it, because that class lives in
:base-services, i.e. in the distribution this build produces rather than
in the build logic that produces it.

AbstractVersionsUpdateTask therefore grew its own hardened factory, which
is in fact stricter than XmlFactories: besides FEATURE_SECURE_PROCESSING
it disallows doctype declarations outright and denies access to external
general entities, parameter entities, DTDs and schemas. That is the
behaviour build logic should get everywhere it parses XML, so leaving it
as a protected method on one task base class means the next caller either
reaches for a bare DocumentBuilderFactory.newInstance() or copies eight
lines of security-sensitive configuration.

Move it verbatim to gradlebuild:basics, which every build logic module
already depends on, and have AbstractVersionsUpdateTask's only subclass
call site import it directly. No behaviour change: the factory
configuration and the resulting DocumentBuilder are identical.
